requirements.txt
flask==2.3.3
gunicorn==21.2.0
python-telegram-bot==20.6
gspread==5.12.0
oauth2client==4.1.3
requests==2.31.0